FIX: Error Refactoring

+++ src/controllers/comments.controller.js
@@ -1,18 +1,26 @@
 const CommentService = require('../services/comments.service');
 const {
-    InvalidParamsError
+    InvalidParamsError, AuthenticationError
 } = require('../exceptions/index.exception');
 class CommentController {
     constructor() {
         this.commentService = new CommentService();
     }
 
-    createComment = async(req, res, next) => {
+    createComment = async (req, res, next) => {
         try{
             const { userId } = res.locals;
             const { pinId } = req.params;
             let { comment, parentCommentId } = req.body;
             const like = 0;
+
+            if (!userId) {
+                throw new AuthenticationError;
+            }
+
+            if (!pinId || !comment || !parentCommentId) {
+                throw new InvalidParamsError;
+            }
 
             await this.commentService.createComment(
                 userId, pinId, comment, parentCommentId, like
@@ -26,6 +34,11 @@ class CommentController {
     getComment = async(req, res, next) => {
         try{
             const { pinId } = req.params;
+
+            if (!pinId) {
+                throw new InvalidParamsError;
+            }
+
             const commentList = await this.commentService.findAllComment(pinId);
 
             res.status(200).json({ comment : commentList });
@@ -39,6 +52,15 @@ class CommentController {
             const { commentId } = req.params;
             const { comment } = req.body;
             const { userId } = res.locals;
+
+            if (!commentId || !comment) {
+                throw new InvalidParamsError;
+            }
+
+            if (!userId) {
+                throw new AuthenticationError;
+            }
+
             await this.commentService.updateComment(userId, commentId, comment);
 
             res.status(200).json({ message: "댓글이 수정되었습니다." });
@@ -52,6 +74,14 @@ class CommentController {
             const { userId } = res.locals;
             const { commentId } = req.params;
 
+            if (!userId) {
+                throw new AuthenticationError;
+            }
+
+            if (!commentId) {
+                throw new InvalidParamsError;
+            }
+
             await this.commentService.deleteComment(userId, commentId);
 
             res.status(200).json({ message: "댓글이 삭제되었습니다." });
@@ -64,7 +94,15 @@ class CommentController {
         try {
             const { userId } = res.locals;
             const { commentId } = req.params;
-            console.log(commentId);
+            
+            if (!userId) {
+                throw new AuthenticationError;
+            }
+
+            if (!commentId) {
+                throw new InvalidParamsError;
+            }
+
             const result = await this.commentService.likeComment(userId, commentId);
             if(result === 0){
                 res.status(200).json({ message: "댓글 좋아요 취소 성공" });
